Fix editing of the guide
